Remove ScriptScoreQueryBodyFn
From inspection I suspect this case would never be hit, because
ScriptScore doesn't implement the Query trait.
This was the only call site for ScriptScoreQueryBodyFn (which has an odd
name, given it accepted a ScriptScore and not a ScriptScoreQuery).
Therefore I've removed it, and put the ScriptScoreQueryBodyFn2 in its
place
