+++ elastic4s-handlers/src/main/scala/com/sksamuel/elastic4s/handlers/searches/queries/QueryBuilderFn.scala
@@ -55,8 +55,7 @@ object QueryBuilderFn {
     case q: RawQuery => RawQueryBodyFn(q)
     case q: RegexQuery => RegexQueryBodyFn(q)
     case q: ScriptQuery => ScriptQueryBodyFn(q)
-    case q: ScriptScore => ScriptScoreQueryBodyFn(q)
-    case q: ScriptScoreQuery => ScriptScoreQueryBodyFn2(q)
+    case q: ScriptScoreQuery => ScriptScoreQueryBodyFn(q)
     case s: SimpleStringQuery => SimpleStringBodyFn(s)
     case s: SpanContainingQuery => SpanContainingQueryBodyFn(s)
     case s: SpanFirstQuery => SpanFirstQueryBodyFn(s)
